Make Repository to mostly relay on iostream interface
All except open()/close() code of Repository can be implemented in terms
of iostream so they will work with disk based files as well as with
memory based files.

On open(), fail if the underlying file is a memory file;
on close(), do the clean up but skip the file close if it is a memory
file.