Support explicit file grow (not relay on filesystem file holes)
On seek_write_phy, if the seek position ends beyond the underlying file,
the gap/hole is filled with zeros.

This is required to work with memory based files. For disk based files
it may not be necessary.

On linux, a file would have a hole which it would be read as zeros;
other system may implement a fill.

This implementation ensures that the seek will not depend on the details
of the file system.