Impl inline size of 6 bits
Reduced from 8 to 6 bits for encoding the inline size. Also such not
longer is required to be multiple of 2.

The header packed all the bits and size in the first 8 bits leaving the
next 8 bits for storing an extra byte of data (if the inline size is an
odd number) or just padding (zero).

Reference: docs/RFC:-Block-Size-and-Block-Extent-Specification.md
Version 2