ext,test: rm is_more flag
The Segment::load now receive the explicit size (in bytes) of the
segment to read from the file.

Neither the load nor the write will use the is_more flag.

The check of underflow/overflow during the read/write needs to be
reviewed, probably refactored and more tests are needed.