test: workaround for EXPECT_THAT on throws that call func twice
EXPECT_THAT, when a func under test throws an exception, it calls the
func again (twice) which has non-trivial side effects.

This is mostly likely a bug in GoogleTest.

The workaround wraps the func under test and ensure that it is called
once: on the second call, it just rethrow the exception.