alloc: fix closest when no best alloc even exist
If the requested size cannot be allocated neither perfectly nor via
split, try to pick the largest of the chunks.

This probably requires a modification in case that a split exist
but then it fail due a dont_split_fr_threshold. This is still pending.

Add tests to cover this (no perfect, no split case).