alloc,test: chk free list cannot alloc because dont_split_fr_threshold
Test that even with extent that are large enough to fulfill the request,
the allocation fails because the extent cannot be split without leaving
N free blocks where N <= dont_split_fr_threshold.

Added a AllocCoalescedDoesntSplitButCloseSuboptimalHint special test
to document the limitation of the choice of the closest extent hint.