repo,test: simplify out of boundaries rw check
Use Repository's begin_data_blk_nr() and past_end_data_blk_nr() as the
boundaries for the data rw operations (do not assume 1 block of header
only).

Centralize the check in chk_extent_for_rw() and make it work for
full and suballoc extents.

Maintain the semantics that if an extent of 0 block count still has
its block number of of the boundaries, it *is* an ExtentOutOfBounds.