alloc,test: impl TailAllocator
TailAllocator allocs blocks at the end of the repository making it to grow
and deallocs blocks also from the end making it to shrink.

In some sense it is a StackAllocator but without tracking the top of the
stack and because "Stack" may suggest no the alloc/dealloc patterns but
the location of the memory allocated, "Tail" is more appropriated.