segm,test: Segment rw struct methods now work in buffers
The rationale is that we will use IOSegment to read/write
from/into the file and the rest of the API should work
on buffers in memory.

This simplifies the code as only a few boundary checks are needed and
there is no risk of getting a bad file read/write.