mem,test: ensure that a iobase read does not shrink the input buf
If the caller sets a buffer of a size N and it is request to read
M < N, do not resize the buffer (aka shrink). This allows the caller
to prefill part of the buffer without the risk of being deleted
by an unneeded resize.

If M > N do the resize: this prevents an overflow and if the caller
wrote something to the buffer before the reading, the resize may
discard it but it would be override anyways.