segm,test: read segm struct until inline (other end not supp)
The former code read the segment struct consuming all the bytes
explicitly set. This allowed to read segments with and without
inline but the API is old pre-RFC.

The RFC says that a segment is an array of extents and the *count*
of extents is defined somewhere else, not the size (in bytes); and
if no count is given, the segment array ends in a extent in inline
mode.

This commit removes the segment size from the API, breaking a lot
of tests that relay on it but not breaking the ones that expected
an inline as the end of the segment.