rfc: rm object descriptor and object concept
Deleted the concept of object descriptor. Only distinguish between
descriptors that owns a segment (and therefore data external to the
descriptor itself) and non-owners. This replaces the `is_obj` and
`obj_id` basically.

Also, if the descriptor requires an explicit persistent id does not
depend of if the descriptor owns or not a segment (so owns_segm does not
imply has_id). This means that `type` now it is always of 0 bits.