ext,io,repo,test: replace null by empty extent
The null extent was used for two different contexts: it was an extent
that pointed to the blk 0 and it was an extent of 0 blks.

Moreover the blk nr 0 was considered forbidden while in reality is is
only forbidden for reading/writing the Repository. In any other context
the blk nr 0 should be like any other blk.

Removed the is_null and replaced by is_empty to make the meaning
explicit and fixed the places where it was assumed that blk nr 0 is bad
or is_null implies blk nr 0.