segm,test: don't allow segment extension if has inline
The rationale is that inline data means to be at the end of the segment
so once a segment has inline data, it cannot be extended.
Caller needs to get a copy of it, remove it, extend the segment
and readd the inline. It is annoying but makes explicit the point.