alloc,demo,dsc,ext,test: make BlockArray to own a SegmentAllocator
The block array now hold a SegmentAllocator object. This requires
the SegmentAllocator to be initialized in 2 steps: first the constructor
with a nullptr reference to the BlockArray and then an assignation
of the BlockArray and its parameters.

This 2-step construction allows subclasses of BlockArray (like
Repository) to create the SegmentAllocator first, load from somewhere
the blk sz and other params, and only then complete the initialization
of the SegmentAllocator.