blk: block array based on a segment
SegmentBlockArray implements the BlockArray interface but instead
of being backed by a file (like Repository), it is backed by
a Segment (and IOSegment). This allows to see a piece of a file
(a segment) and do suballoctions (as if it were a "sub" block array).

The grow/shrink logic aims to be simple, only taking some care
for not generate too many split/too many tiny extents (and therefore
to large segment footprint). Still, it is caller responsibility to
do cleaver grow/shrink strategies.