blk: fix chks (small sz, power of 2 and _real update)
The blk size can be arbitrary small, and if it is too small, the
suballocation will be disabled (instead of throwing).

A new requirement is that the size is a power of 2. This comes from
several assumptions in Extent and SegmentAllocator.

Also remove an assert on _real_past_end_blk_nr that didn't make sense.