blk,repo: read/write bytes at blk nr + offset, not by extent
Make BlockArray public read_extent and write_extent to call impl_read
and impl_write (subclass). This makes subclasses of BlockArray to work
at a block number plus offset and in bytes and not in terms of extent.
This enables BlockArray to handle the full-allocated/sub-allocated
cases like Repository does.