blk: avoid full read/write when rd/wr suballoc extents
The initial implementation read the entire block and then read/write
the selected subblocks of the extent. This is inefficient if the rd/wr
is for a single subblock (has 16 times the cost).

The new implementation only reads/writes what is absolutely needed. This
increases the number of seeks and going to the disk for the multiple
small read/write ops. We relay in the subclass implementation and let
it to handle how to buffer this. (The libc does this for files).