segm,test: explicit end mode for read/load segment
This generalize the previous implementation and adds the possibility
to end a segment when the io ends. This mode allows the caller to save
a segment in fixed-size places that work as natura boundaries without
requiring the user neither store an explicit length or add an
unnecessary inline to mark the end of the segment.