repo,test: add feat flags, chksum, root_sg to repo header
Add a set of feature flags (none in use), a checksum (not computed yet)
and the root segment to the repository header.

This root segment contains the blocks where the descriptor set is
located. The code manages the case if the segment is small enough to be
added to the header or it requires additional space.

Tests where updated accordingly but more tests are needed.