repo,test: make Repository to own a FileBlockArray, not inherit from BlockArray
This is a major refactor making the Repository owner of a FileBlockArray
and not inherit from BlockArray.
This made a lot of code of Repository to disappear and relay on
FileBlockArray's methods for the read/write, grow/shrink and block
management.

Only a small _grow_by_blocks/_shrink_by_blocks API is exposed only for
testing.

The code still has a few TODOs and code that should be removed
or cleaned up.