dsc: impl checksum and header for descriptor set
Any descriptor set, even empty ones, has an header of 4 bytes
with the inet checksum of the set.

The set now has a create_set() method to create a fresh empty
set (with a fresh header) in addition to the old load_set() method
that loads the set and the header from the underlying file.