dsc: defer external blk deallocation to final descriptor erasing
If the segment of a to-be-deleted descriptors are deallocated but
the descriptor is not zeroed, it may be possible to another descriptor
alloc those same blocks and a crash then happen. The xoz file will
then have two descriptors pointing to the same segment.
If we zeroed the former, such case should not happen.