dsc,test: preserve dsc on failed move_out(); make caller chk before add()
Due a limitation of C++'s unique_ptr<>, we cannot pass it
to add() to do the checks without having to give the ownership
(this happen when the unique_ptr<> points to a parent/base class).

Because of this, the caller will have to explicitly do the check.
add() will still do the same but if an error happen, the descriptor
will be lost.

For move_out() we can do the full check before touching the descriptors
so it is safer.