dsc: subclass must update_header(), supersedes pre_write_struct()
The reasoning is that if the pre_write_struct() is the only moment to
get updated the header (hdr), this is too late if the caller wants
to know the descriptor's footprint (calc_struct_footprint_size
or calc_data_space_size) *before* doing a call to write_struct_into().

The update_header() is for this and supersedes pre_write_struct().