dsc,test: make DescriptorSet to own a priv segment, no longer a ref
Having a reference to a typically short-lived object like Segment
was a bad idea just to avoid the copy of a few bytes.
Now, DescriptorSet will have a private copy of the argument
that will modify as it discretion.
Caller may get a updated segment from the set calling segment() method.