alloc,blk,dsc,ext,io: tag as "unreachable"/unused
Some code is only for consistency checks that are disabled
when NDEBUG is set. Therefore, the compiler complains of
unused variables or missing return in functions.
These can be silenced by using [[maybe_unused]] and std::terminate()

Note: in C++23 there is a [[unreachable]] and it should be more
efficient than calling std::terminate() because the compiler will
not emit a func call and instead assume that the branch will be
never taken.