alloc,blk,demo,dsc,io,repo,segm,test: make Segment blk_sz_order aware
While in theory a Segment is independent of the block array, in practice
each segment "points" to the same block array all of its lifetime.
Because calculating the data space size requires the blk_sz_order,
having it as an attribute is convenient at the expense of wasting
a few extra bytes.

When comparing two segments, it is an error to compare two
with different blk_sz_order because that would mean to compare
two locations of different blk arrays (and makes little sense).