dsc: impl release_free_space() and flush_writes() for desc
Because a descriptor may like to defer the writes to disk or
defer the deallocation of unused space (like dset_holder does),
the methods are not exposed as virtual with a no-op default
implementation.

The full_sync() method simplifies the pattern of calling
flush_writes() followed by update_header() (to get an accurate
size) or with the combination of release_free_space() (to get also
the minimum size of the file).