repo: handle the case of trampoline segment too large
This is an unlikely case where the trampoline segment gets
so fragmented that it is too large to fit in the header.
We are not going for a "trampoline of a trampoline" solution,
instead we discard the trampoline segment and allocates a single
extent, not matter if this is less space efficient.