dsc: chk that segment content and set content agree on checksum
We detected a bug of uninitialized data with this check.
On one hand the set calculates the checksum based on the
checksum of its descriptors; on the other hand, we compute
the checksum of the entire set's segment.
In theory both checksum should agree but it was not the case
due an overallocation (the segment grew more than expected) and
this unexpected additional space was never zeroed, hence
the checksum failed.