dsc,test: zero'd unused allocated space on dset's segment
When allocating new data blocks for the descriptor set at the
segment-block-array level, the allocator may allocate more space
(taken from the underlying block array).
This extra space remains inaccessible from the set (blkarr capacity)
but by default this space is also uninitialized.

On load the set, the code process all the content of the
segment-block-array without knowing anything about the blkarr's capacity
so it will eventually read this extra uninitialized space.

To fix this on write, the set will bypass the segment-block-array
and it will fill with 0x00 any extra newly added space in the
segment's space. On load, the set will interpret these 0x00 as padding
and it will not change neither the set composition nor the checksum.