dsc,repo: mv dsc mapping into runtime ctx; let be a init param of Repo
The use of a global (static) mapping table does not scale and makes
impossible to have two different kind (or even version) of xoz files
with different mappings in the same binary.
Moving the descriptor mapping to the runtime context fixes this.

To make the Repository configurable, it will receive such runtime
context via parameter.