dsc: preserve idata from future
On reading/loading, a descriptor may find idata beyond what it was
expecting. This data may come from a future version of the App
(or put in other words, a future App wrote a xoz file and now we
are reading it with and older App version).

This data from the future must be preserved, otherwise, writing
with an older App destroys what a newer App wrote and this breaks
future/forward compatibility.

The mechanism implemented covers only idata: cdata is not covered
and it is (for now, app's job).
Note also that this mechanism does not guarantees forward compatibility:
it is also needed the help of the app.