dsc,test: dset's sg blkarr realloc on grow by default
Make the descriptor set's SegmentBlockArray to do reallocations
on grow requests. This is much more efficient than previous setup
when the grows are of tiny sizes like the ones that we expect
to find in DescriptorSet.

The tests for DescriptorSet were updated to validate this new default
(expected a reduced footprint in disk); the tests for File were left
using the previous runtime config to avoid rewriting all the tests
again.