dsc,test: full_sync each dsc before dset flush writes
To handle defer/pending writes on each descriptor properly,
we need to do a full_sync of them and only then do a flush_writes
in the set, even if the caller wanted a flush_writes of the set only.

In practice, no caller will do a flush_writes alone because it is
a private method of Descriptor and it is much easier to think
in terms of recursive full_sync.