file,test: load the id_mapping dsc (wip)
Load the IDMappingDescriptor from the root set if it is present,
otherwise create a default and add it to the set.

If add_missing_id_mapping_to_root_set is false, do not add it
(this is required to not break existing tests).

Currently the loaded descriptor is not used in this commit
but it will in the next ones.